Fix off-by-7 error when getting activity data
We were previously getting data for selectedDate, which is the starting
date for the period the user selected. This meant that data for the
current week would not be included in the set.

Also changed userId in the URL to just '-', since that is what Fitbit
suggests for the current user: https://dev.fitbit.com/docs/activity/